requirements.txt
PyYAML==6.0.1
watchdog==3.0.0
